refactor values
