+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/Base64GzipFunction.cs
@@ -20,6 +20,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Text;
 using System.Text.Json;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
