+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/EvaluateStartEndWith.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Collections.Generic;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
