+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/JoinFunction.cs
@@ -16,6 +16,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
