+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/JsonFunction.cs
@@ -15,6 +15,7 @@
 
 using System.Collections.Generic;
 using System.Diagnostics;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
