+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Functions/ToNumberFunction.cs
@@ -16,6 +16,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath.Functions;
