+++ libraries/src/AWS.Lambda.Powertools.JMESPath/JsonTransformer.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Text.Json;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath
 {
