+++ libraries/src/AWS.Lambda.Powertools.JMESPath/Token.cs
@@ -16,6 +16,7 @@
 using System;
 using System.Diagnostics;
 using AWS.Lambda.Powertools.JMESPath.Functions;
+using AWS.Lambda.Powertools.JMESPath.Values;
 
 namespace AWS.Lambda.Powertools.JMESPath
 {
